pyproject.toml
[tool.poetry]
name = "mkdocs-multirepo"
version = "0.2.0"
description = ""
authors = ["Lars Wilhelmer"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3"
beautifulsoup4 = "^4.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
mkdocs-multirepo = 'mkdocs_multirepo:main'